Cleanup (#125)
* added max size for batching messages

* default message batch size

* wip: chaos.

* Add benchmark actor and protobuf message

The update introduces a benchmark actor that receives messages and increments a message counter. It also includes the creation of a new protobuf message. The main function is updated to incorporate this benchmark actor, while providing a detailed simulation of sending messages across multiple actors distributed across different engines.

* Refactor benchmark code and add profiling

The benchmark code is refactored for better error handling and message checks. 'benchmark' function is extracted from 'main' for further testing. A new Makefile target 'bench-profile' is added for profiling. A new test file 'main_test.go' is created for benchmark testing. Corresponding instructions are added in the newly created README.md file. Also, .gitignore is updated to exclude the created test, cpu, and memory profiles.

* document how to use the interactive web interface.

* add latency measurement as well.

* bah. can't make go test ignore the new latency tests. I just commented it out. Gonna revisit latency benchmarks later.

* Update benchmark command in Makefile

The benchmark command in the Makefile has been updated to run the whole package, not just main.go

* misc cleanup

- move various log messages from Info to Debug, so they won't show up.
- remove unused errors.go
- add test for duplicate actor Event
- rename EngineOpts to EngineConfig so it matches the rest of the codebase.
 - a few questions that are no longer questions have been removed

* naming consistency.

- name ==> kind
- id ==> kind

* formatting

* Opts --> config

* cluster.Start() doesn't need to return error.

---------

Co-authored-by: anthdm <cryptoanthdm@gmail.com>